% Auto-generated: do not edit by hand
\name{moonThreejs}

\alias{moonThreejs}

\title{MoonThreejs component}

\description{

}

\usage{
moonThreejs(id=NULL, height=NULL, nforests=NULL, ntrees=NULL,
radius=NULL)
}

\arguments{
\item{id}{Character. }

\item{height}{Numeric. }

\item{nforests}{Numeric. }

\item{ntrees}{Numeric. }

\item{radius}{Numeric. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

